added speend

+++ main.py
@@ -4,6 +4,7 @@ from fastapi.responses import FileResponse
 from contextlib import asynccontextmanager
 from env import db_credentials
 from fastapi import FastAPI
+import pandas as pd
 import numpy as np
 import asyncio
 import os
@@ -61,17 +62,30 @@ def index():
 # Endpoint to get the DataFrames
 @app.get("/data/trade/jp/")
 async def get_data(time:str, types:str, agr:bool=False, group:bool=False):
-    df = dt.process_int_jp(time, types, agr, group)
-    return df.to_pandas().to_dict()
+    if not os.path.exists(os.path.join(os.getcwd(), "data", "processed", f"jp_{time}_{types}.parquet")):
+        df = dt.process_int_jp(time, types, agr, group)
+        df.to_parquet(os.path.join(os.getcwd(), "data", "processed", f"jp_{time}_{types}.parquet"))
+        return df.to_pandas().to_dict()
+    else:
+        return pd.read_parquet(os.path.join(os.getcwd(), "data", "processed", f"jp_{time}_{types}.parquet")).to_dict()
 
 @app.get("/data/trade/org/")
 async def get_org_data(time:str, types:str, agr:bool=False, group:bool=False):
-    df = dt.process_int_org(time, types, agr, group)
-    return df.to_pandas().to_dict()
+    if not os.path.exists(os.path.join(os.getcwd(), "data", "processed", f"org_{time}_{types}.parquet")):
+        df = dt.process_int_org(time, types, agr, group)
+        df.to_parquet(os.path.join(os.getcwd(), "data", "processed", f"org_{time}_{types}.parquet"))
+        return df.to_pandas().to_dict()
+    else:
+        return pd.read_parquet(os.path.join(os.getcwd(), "data", "processed", f"org_{time}_{types}.parquet")).to_dict()
 
 @app.get("/data/trade/moving/")
 async def get_moving_data(agr:bool=False):
-    return dt.process_price(agr=agr).to_pandas().replace([np.nan, np.inf, -np.inf], [0, 0, 0]).to_dict()
+    if not os.path.exists(os.path.join(os.getcwd(), "data", "processed", "moving.parquet")):
+        df = dt.process_price(agr=agr)
+        df.to_parquet(os.path.join(os.getcwd(), "data", "processed", "moving.parquet"))
+        return df.to_pandas().replace([np.nan, np.inf, -np.inf], [0, 0, 0]).to_dict()
+    else:
+        return pd.read_parquet(os.path.join(os.getcwd(), "data", "processed", "moving.parquet")).replace([np.nan, np.inf, -np.inf], [0, 0, 0]).to_dict()
 
 @app.get("/data/index/consumer")
 async def get_consumer(update:bool=False):
@@ -91,16 +105,20 @@ async def get_trade_file(time:str, types:str, agr:bool=False, group:bool=False):
 
 @app.get("/files/trade/org/")
 async def get_org_file(time:str, types:str, agr:bool=False, group:bool=False):
-    df = dt.process_int_org(time, types, agr, group)
-    file_path = os.path.join(os.getcwd(), "data", f"{time}_{types}.csv")
-    df.to_csv(file_path)
+    file_path = os.path.join(os.getcwd(), "data", "processed", f"org_{time}_{types}.csv")
+    
+    if not os.path.exists(os.path.join(os.getcwd(), "data", "processed", f"org_{time}_{types}.csv")):
+        df = dt.process_int_org(time, types, agr, group)
+        df.to_csv(file_path)
     return FileResponse(file_path, media_type='text/csv', filename=f"{time}_{types}.csv")
 
 @app.get("/files/trade/moving")
 async def get_moving_file(agr:bool=False):
-    df = dt.process_price(agr=agr)
-    file_path = os.path.join(os.getcwd(), "data", "moving.csv")
-    df.to_csv(file_path)
+    file_path = os.path.join(os.getcwd(), "data", "processed", "moving.csv")
+    
+    if not os.path.exists(os.path.join(os.getcwd(), "data", "processed", "moving.csv")):
+        df = dt.process_price(agr=agr)
+        df.to_csv(file_path)
     return FileResponse(file_path, media_type='text/csv', filename="moving.csv")
 
 @app.get("/files/index/consumer")
